Add Makefile to help create `requirements.txt`
Rather than installing poetry in the CI/CD pipeline, we'll just install
dependencies using a `requirements.txt` file. This introduces a bit of
manual work whenever new dependencies are added to the project, but
since this is a rather uncommon event, it's worth the gain in
performance during the build stage in the CI/CD pipeline.

We'll see if we can further automate this to prevent any unnecessary
issues whenever someone forgets to update the `requirements.txt`
file (e.g., using a post-install hook in poetry --if such a thing was
supported-- would be ideal; otherwise, we'd need to shim package
installation to do both things as a single action)